Update your cryptocurrency application from the spring-testing module:
Update your "bitcoin" endpoint so that it can take the name of the cryptocurrency to get information about in its request.
Update all your tests accordingly to ensure that the currency you're getting information about is the currency that was requested.
Add httpBasic() authentication to your new endpoint.
Update your tests accordingly.